requirements.txt
pydantic>=2.7
python-dotenv>=1.0
requests>=2.31
pandas>=2.2
matplotlib>=3.8
geopandas>=0.14
shapely>=2.0
jinja2>=3.1
rich>=13.7
loguru>=0.7
python-docx>=1.1.0
